Add files via upload
Creates the following folders within the working directory that the shell is run in: "lib", "spec", and "spec/lib". The code asks the user for the name of the Ruby file they want to create within that directory and creates a Ruby file with that name within the "lib" folder and its associated Spec file within the "spec/lib" folder. The names are formatted as follows: If user input is equal to file, the file within the "lib" folder will be named "file.rb" and the file within the "spec/lib" folder will be named "file_spec.rb.